cloud-function/requirements.txt
functions-framework
fpdf2
pymupdf
google-cloud-storage
google-cloud-aiplatform>=1.38.0

